requirements.txt
Flask==3.0.3
flask-cors==4.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
gunicorn==21.2.0
Werkzeug==3.0.3